+++ src/main/java/gov/nasa/jpl/edrn/dmcc/auth/DMCCAuthenticationInterceptor.java
@@ -4,19 +4,19 @@
 package gov.nasa.jpl.edrn.dmcc.auth;
 
 // Apache™ DS
-import org.apache.directory.server.core.CoreSession;
-import org.apache.directory.server.core.DefaultCoreSession;
-import org.apache.directory.server.core.DirectoryService;
-import org.apache.directory.server.core.interceptor.BaseInterceptor;
-import org.apache.directory.server.core.interceptor.context.BindOperationContext;
-import org.apache.directory.server.core.interceptor.NextInterceptor;
-import org.apache.directory.server.core.LdapPrincipal;
+import org.apache.directory.server.core.api.CoreSession;
+import org.apache.directory.server.core.api.DirectoryService;
+import org.apache.directory.server.core.api.interceptor.BaseInterceptor;
+import org.apache.directory.server.core.api.interceptor.context.BindOperationContext;
+import org.apache.directory.server.core.api.LdapPrincipal;
+import org.apache.directory.server.core.shared.DefaultCoreSession;
 
 // Apache™ Shared LDAP API
-import org.apache.directory.shared.ldap.constants.AuthenticationLevel;
-import org.apache.directory.shared.ldap.name.DN;
-import org.apache.directory.shared.ldap.name.RDN;
-import org.apache.directory.shared.ldap.util.StringTools;
+import org.apache.directory.shared.ldap.model.constants.AuthenticationLevel;
+import org.apache.directory.shared.ldap.model.exception.LdapException;
+import org.apache.directory.shared.ldap.model.name.Dn;
+import org.apache.directory.shared.ldap.model.name.Rdn;
+import org.apache.directory.shared.util.Strings;
 
 // DMCCrap
 import org.fhcrc.perdy.edrn_ws.ws_newcompass.WsNewcompass;
@@ -28,6 +28,7 @@ import org.slf4j.LoggerFactory;
 
 // Misc
 import java.net.URL;
+import java.net.MalformedURLException;
 
 /**
  * An interceptor that delegates authentication to the DMCC authentication web service,
@@ -106,18 +107,29 @@ public class DMCCAuthenticationInterceptor extends BaseInterceptor {
     }
 
     /**
+     * Construct a DMCCAuthenticationInterceptor.
+     */
+    public DMCCAuthenticationInterceptor() {
+        super("dmccAuthenticationInterceptor");
+    }
+
+    /**
      * The initialization method that … uhm … 
      */
-    public void init(DirectoryService directoryService) throws Exception {
+    public void init(DirectoryService directoryService) throws LdapException {
         super.init(directoryService);
-        this.directoryService = directoryService;
-        this.soapService = new WsNewcompass(new URL(this.authWebServiceURL)); // DMCC's service name is stupid.
-        LOG.debug("Created SOAP service {}", this.soapService);
-        StringBuffer b = new StringBuffer(this.checkDigitLength);
-        for (int i = 0; i < this.checkDigitLength; ++i)
-            b.append('0');
-        this.checkDigits = b.toString();
-        LOG.debug("Created {} zeros for the check digits (ugh)", this.checkDigitLength);
+        try {
+            this.directoryService = directoryService;
+            this.soapService = new WsNewcompass(new URL(this.authWebServiceURL)); // DMCC's service name is stupid.
+            LOG.debug("Created SOAP service {}", this.soapService);
+            StringBuffer b = new StringBuffer(this.checkDigitLength);
+            for (int i = 0; i < this.checkDigitLength; ++i)
+                b.append('0');
+            this.checkDigits = b.toString();
+            LOG.debug("Created {} zeros for the check digits (ugh)", this.checkDigitLength);
+        } catch (MalformedURLException ex) {
+            throw new LdapException(ex);
+        }
     }
 
     /**
@@ -127,7 +139,7 @@ public class DMCCAuthenticationInterceptor extends BaseInterceptor {
      * @param password The password for the username.
      * @return True if authentic and valid, false if not.
      */
-    public boolean isAuthentic(String username, String password) throws Exception {
+    public boolean isAuthentic(String username, String password) {
         WsNewcompassSoap port = this.soapService.getWsNewcompassSoap12(); // DMCC's port names are stupid.
         String response = port.pwdVerification(username, password, this.checkDigits);
         LOG.debug("DMCC says that for username {} and password [REDACTED] the response is {}", username, response);
@@ -138,40 +150,37 @@ public class DMCCAuthenticationInterceptor extends BaseInterceptor {
      * Intercept the "bind" call and attempt to validate the password with the DMCC.  If success,
      * no need to continue the chain.  If not, then perhaps something else in the chain can do it.
      *
-     * @param next Next interceptor in the chain.
-     * @param opContext Context for the bind operation.
+     * @param bindContext Context for the bind operation.
      * @throws Exception Should something—anything—go wrong.
      */
     @Override
-    public void bind(NextInterceptor next, BindOperationContext opContext) throws Exception {
+    public void bind(BindOperationContext bindContext) throws LdapException {
         // TODO: to be really anal retentive, we should check the entire DN and ensure it makes sense
         // for the DMCC to authenticate, on the off chance that "uid=dcrichto,ou=Sales,o=Amazon,c=US"
         // happens to authenticate with our LDAP server.
         try {
             LOG.debug("★ In DMCCAuthenticationInterceptor's bind");
-            DN dn = opContext.getDn();
-            RDN rdn = dn.getRdn();
+            Dn dn = bindContext.getDn();
+            Rdn rdn = dn.getRdn();
             if (!rdn.getNormType().equals(this.rdnAttributeType))
                 throw new DMCCAuthenticationException("Can't handle RDN attribute type " + rdn.getNormType() + "; expected "
                     + this.rdnAttributeType);
-            String uid = rdn.getNormValue();
-            String password = StringTools.utf8ToString(opContext.getCredentials());
+            String uid = rdn.getNormValue().getString();
+            String password = Strings.utf8ToString(bindContext.getCredentials());
             LOG.debug("Got uid '{}' with password 'NOPE! Not gonna show ya!'", uid);
             if (!this.isAuthentic(uid, password))
                 throw new DMCCAuthenticationException("Uid '" + uid + "' and password not valid per DMCC");
             LOG.debug("Guess what: your DMCC username & password are OK!");
-            LdapPrincipal principal = new LdapPrincipal(opContext.getDn(), AuthenticationLevel.SIMPLE);
+            LdapPrincipal principal = new LdapPrincipal(this.schemaManager, bindContext.getDn(), AuthenticationLevel.SIMPLE);
             CoreSession session = new DefaultCoreSession(principal, this.directoryService);
-            opContext.setSession(session);
-            opContext.setCredentials(null);
+            bindContext.setSession(session);
+            bindContext.setCredentials(null);
         } catch (RuntimeException ex) {
             throw ex;
         } catch (Exception ex) {
             // Can't authenticate via the DMCC.  Let another interceptor try.
             LOG.debug("Unable to authenticate with DMCC: {}, {}", ex.getClass().getName(), ex.getMessage());
-            next.bind(opContext);
+            next(bindContext);
         }
     }
-    
-    
 }